Refactor notification media
