+++ src/argus/notificationprofile/media/__init__.py
@@ -1,64 +1,26 @@
-import json
 import logging
-from abc import ABC, abstractmethod
 from multiprocessing import Process
 from typing import List
 
 from django.conf import settings
 from django.db import connections
-from django.template.loader import render_to_string
-from rest_framework.renderers import JSONRenderer
 
 from argus.auth.models import User
 from argus.incident.models import Incident
-from argus.incident.serializers import IncidentSerializer
-from .models import NotificationProfile
+
+from ..models import NotificationProfile
+from .email import EmailNotification
 
 
 LOG = logging.getLogger(__name__)
 
 
-class NotificationMedium(ABC):
-    @staticmethod
-    @abstractmethod
-    def send(incident: Incident, user: User):
-        pass
-
-
-class EmailNotification(NotificationMedium):
-    @staticmethod
-    def send(incident, user):
-        if not user.email:
-            logging.getLogger("django.request").warning(
-                f"Cannot send email notification to user '{user}', as they have not set an email address."
-            )
-
-        title = f"Incident at {incident}"
-        incident_dict = IncidentSerializer(incident, context={IncidentSerializer.NO_PKS_KEY: True}).data
-        # Convert OrderedDicts to dicts
-        incident_dict = json.loads(JSONRenderer().render(incident_dict))
-
-        template_context = {
-            "title": title,
-            "incident_dict": incident_dict,
-            "longest_field_name_length": len(max(incident_dict, key=len)),
-        }
-        try:
-            user.email_user(
-                subject=f"{settings.NOTIFICATION_SUBJECT_PREFIX}{title}",
-                message=render_to_string("notificationprofile/email.txt", template_context),
-                html_message=render_to_string("notificationprofile/email.html", template_context),
-            )
-        except ConnectionRefusedError as e:
-            EMAIL_HOST = getattr(settings, "EMAIL_HOST", None)
-            if not EMAIL_HOST:
-                LOG.error("Notification: Email: EMAIL_HOST not set, cannot send")
-            EMAIL_PORT = getattr(settings, "EMAIL_PORT", None)
-            if not EMAIL_PORT:
-                LOG.error("Notification: Email: EMAIL_PORT not set, cannot send")
-            if EMAIL_HOST and EMAIL_PORT:
-                LOG.error('Notification: Email: Connection refused to "%s", port "%s"', EMAIL_HOST, EMAIL_PORT)
-            # TODO: Store error as incident
+__all__ = [
+    "send_notifications_to_users",
+    "background_send_notifications_to_users",
+    "send_notification",
+    "get_notification_media",
+]
 
 
 MODEL_REPRESENTATION_TO_CLASS = {
