+++ src/argus/notificationprofile/media/base.py
@@ -1,0 +1,14 @@
+from abc import ABC, abstractmethod
+
+from argus.auth.models import User
+from argus.incident.models import Incident
+
+
+__all__ = ['NotificationMedium']
+
+
+class NotificationMedium(ABC):
+    @staticmethod
+    @abstractmethod
+    def send(incident: Incident, user: User):
+        pass
